snap-group: Consolidate snap ratio and window bounds calculation
This cl consolidated snap ratio calculation between snap group and
`SplitViewController` so that the divisor for the snap ratio calculation
will be the work area subtract the divider shorter side width if exists.

The graph to illustrate the idea can be found at:
http://b/322576687#comment2

Fixed: b/322576687
Test: Modify existing
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5314246
Commit-Queue: Michele Fan <michelefan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265541}

+++ ash/wm/window_state.cc
@@ -255,10 +255,10 @@ float AdjustCurrentSnapRatio(aura::Window* window,
       ShouldConsiderDivider(window) ? kSplitviewDividerShortSideLength / 2 : 0;
   if (IsLayoutHorizontal(window)) {
     return static_cast<float>(target_bounds.width() + divider_delta) /
-           static_cast<float>(maximized_bounds.width());
+           maximized_bounds.width();
   }
   return static_cast<float>(target_bounds.height() + divider_delta) /
-         static_cast<float>(maximized_bounds.height());
+         maximized_bounds.height();
 }
 
 // Move all transient children to |dst_root|, including the ones in the child
